refactor(cli): the Decision Log gets its own reader, under the size ceiling
`scripts/lib/product-lint.js` reached 703 lines — over the 700-line hard limit
in `.claude/rules/coding-standards.md`, which `product-roadmap.js`'s own
docblock cites as the reason for the previous split. Nothing counts lines in
CI, so lint, `npm test` and all six checks passed green over a file that
breached the standard.

Third crossing, same cut as the second: one on-disk format, one owner.
`scripts/lib/product-decisions.js` takes the Decision Log's reader —
`parseDecisions`, `checkDecisionNumbering`, `checkStatusPresence`,
`statusClauses`, `checkSupersededStatus`, `checkFlipsAreClaimed`,
`checkRelations` and the log's constants — so the rules about the log's own
shape (C2's numbering invariants, C3's relation and status coherence) travel
with the parser, the way C4's arity check and C6's framing clause travel with
the roadmap table's. `product-lint.js` comes out at 400 and keeps the rules
that read the parsed state: C1, C4's row↔spec pairing, C5, C6's sanity floor,
C7 and `validateProduct`.

Mechanical — no rule moved and no behaviour changed. The proof is the suite:
`check-product.test.js` requires only `validateProduct`, so a green run across
the move shows the rules run as they did (1987 tests, all six checks green).

`scripts/check-product.js` imports `parseDecisions` from the new module
directly and `product-lint.js` stops re-exporting it, per the no-barrel rule.
Prose touched: the docblock paragraph naming the siblings (two become three,
cut one per format), the cross-module constant names it now qualifies, and
`check-product.js`'s own header.

`docs/plans/check-product-deferred.md` §4 records the crossing with the landed
line counts, and states the standing constraint plainly: nothing mechanically
counts lines, so a size lint is what would close it — deferred to the
controller because it needs a grandfathered-file allowlist.
