+++ scripts/check-product.js
@@ -6,10 +6,13 @@
  * Reads the hand-maintained product state under `product/` — the roadmap table
  * and Decision Log in `ROADMAP.md`, plus every living spec in `specs/` — and
  * validates it with scripts/lib/product-lint.js, which owns the C1–C7 rules
- * (reading the roadmap table through scripts/lib/product-roadmap.js)
  * (the `← we are here` marker, the log's numbering and supersession
  * invariants, spec headers against their governing roadmap row, spec citations,
- * the sanity floor, and the `Tag` cell). Fits the scripts/check-*.js family.
+ * the sanity floor, and the `Tag` cell). Each of the two formats in `ROADMAP.md`
+ * has its own reader beside it — the roadmap table's is
+ * scripts/lib/product-roadmap.js, the Decision Log's is
+ * scripts/lib/product-decisions.js — and this file imports from each directly
+ * rather than through a re-export. Fits the scripts/check-*.js family.
  *
  * CLI tier: prints a report and exits 0 (valid) / 1 (invalid).
  */
@@ -17,8 +20,9 @@
 const fs = require('node:fs');
 const path = require('node:path');
 
-const { validateProduct, parseDecisions } = require('./lib/product-lint');
+const { validateProduct } = require('./lib/product-lint');
 const { parseRoadmapRows } = require('./lib/product-roadmap');
+const { parseDecisions } = require('./lib/product-decisions');
 
 const PRODUCT_DIR = path.resolve(__dirname, '..', 'product');
 const ROADMAP_MD = path.join(PRODUCT_DIR, 'ROADMAP.md');
